Create Menu Design Notes
Working on a menu system. Two different menus will be used with a stack. One menu selects a queue to edit, the other menu cycles through queue items and allows editing on the current item. There will be 3 queues: a Today queue, a Tomorrow queue, a Someday queue. Priority values will be specified based on itemCount, not user preference. The user will select a queue to add items to. Items will be displayed on a first in first out basis unless they are skipped.